requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pytesseract==0.3.10
pdf2image==1.16.0
pillow==10.0.0
motor==3.4.0
pymongo==4.6.0

python-multipart==0.0.6

regex==2025.11.3

python-dotenv==1.0.0
